[optimize] moved connection from conn to transport
